now catalog extend ctentry so it can have attribute and localized descriptions

+++ serpics/core/src/main/java/com/serpics/core/datatype/ProductType.java
@@ -1,10 +1,10 @@
 package com.serpics.core.datatype;
 
 public final class ProductType {
-	public final static short ITEM = 1;
-	public final static short PRODUCT = 2;
-	public final static short BUNDLE = 3;
-	public final static short PACKAGE = 4;
-	public final static short SERVICE = 5;
+	public final static Integer ITEM = 1;
+	public final static Integer PRODUCT = 2;
+	public final static Integer BUNDLE = 3;
+	public final static Integer PACKAGE = 4;
+	public final static Integer SERVICE = 5;
 
 }
